backend/requirements.txt
flask
flask-socketio
eventlet
opencv-python
numpy
python-dotenv
requests
matplotlib
tk
fastapi
uvicorn[standard]
SQLAlchemy>=2.0
pymysql>=1.1.0
python-jose[cryptography]
passlib[bcrypt]
pydantic>=2
